chore(models+schemas): rename user related files for clarity
