+++ app/schemas/__init__.py
@@ -1,5 +1,5 @@
 from .job_schemas import JobAppBase, JobAppCreate, JobApp, JobAppUpdate, ApplicationStatus
-from .users import UserRole
+from .user_schemas import UserRole
 
 __all__ = [
     "JobAppBase", "JobAppCreate", "JobApp",
